pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crm-facet-library"
version = "0.1.0"
description = "Shared capability facet library for CRMinventory and CRMBuilder"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Doug Bower" }]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pre-commit>=3.6",
]

[project.scripts]
crm-facet-lint = "crm_facet_library.validation:main"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"crm_facet_library" = ["LIBRARY.yaml", "facets/**/*.yaml"]
